Skunk Program - Simulator Runs Fine
Program for the Skunk flowmeter controller, currently runs to take input
from and send output to a terminal (or rather command prompt). Some
improvements can be made. Once the some functions are redefined to run
on the arduino (which will involve some lab time), then this program
will be able to control the flowmeter (hopefully!). (Also, some
improvements do need to be made still, e.g. actual settings to change.)