Significant updates - see extended description
+ Followed some suggestions concerning defining arduino or pc at start
of program, to allow certain parts to compile depending on situation -
Many thanks to Adam and David.
+ Added in serLCD libraries to same folder and interatctions in display
function.
+ Shortened all strings to only output max 16 chars, and added double
checking on display, and at other places
+ Some restructuring of various functions for display improvements (no
redraws if not needed etc)
+ Backlight setting added
+ Arduino updateGlobalTimer now deals with the overflow of millis() with
a minor pause, this could take 50 days to occur, and its unlikely that
its gonna be left on that long, but better safe than sorry